requirements.txt
Flask
Pillow
gevent
# gunicorn==19.9.0
keras
tensorflow
numpy
